requirements.txt
keras
pillow
tk
numpy
